osd: do not remove old target before renaming recovered object
The collection_move_rename operation now does an atomic replacement.

In particular, we want to avoid a situation where we

1- create temp object
2- write to it
3- write more to it
4- remove target
5- rename temp object over target

and fail shortly after.  If the journal replays from 2, we won't have the
temp object, and step 5 will delete the target without providing a
replacement at all... losing both the old and new version.

Signed-off-by: Sage Weil <sage@redhat.com>